feat: add navigation and split recommendations
